removed LogicFactory, renamed RandomLogic to Logic, removed strategie parameter
