--- game_players/hase_und_igel_player_new/src/sc/player2018/logic/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Dieses package enthaelt die Spieler-Logiken. <br/>
- * Anfangs ist dies nur die {@link sc.player2018.logic.RandomLogic Zufalls-Strategie}. <br/>
- * fuer umfangreichere Logiken mit vielen Klassen koennte man 
- * hier Unterpackages anlegen, z.B. "sc.player.logic.logicA"
- */
-package sc.player2018.logic;
