+++ game_players/hase_und_igel_player_new/src/sc/player2018/package-info.java
@@ -1,5 +1,5 @@
 /**
- * Dieses package enthaelt die Spieler-Logiken und Verwaltungsklassen
- * fuer diese. <br/> Die Hauptklasse des Simpleclients ist {@link sc.player2017.Starter}.
+ * Dieses package enthaelt die Spieler-Logik.
+ * <br/> Die Hauptklasse des Simpleclients ist {@link sc.player2018.Starter}.
  */
 package sc.player2018;
